EZ3062- update types of new table ID's

+++ Backend/ModelsWithDB/Alibaba/AlibabaBuyer.cs
@@ -12,7 +12,7 @@
         [FK("Customer", "Id")]
         public int? CustomerId { get; set; }
         [FK("AlibabaContract", "ContractId")]
-        public long? ContractId { get; set; }
+		public int? ContractId { get; set; }
         [Length(100)]
         public string BussinessName { get; set; }
         [Length(100)]
